requirements.txt
newspaper3k==0.2.8
Requests==2.31.0
streamlit==1.31.0
